refactor: move integration tests of module to Tests folder, temp fix of usings
