+++ backend/src/Tests/FunctionalTests/Tests/AuctionBids/RaiseBid/RaiseBidCommand_Tests.cs
@@ -5,6 +5,8 @@ using Common.Application.Events;
 using Core.Command.Bid;
 using Core.Common.Domain;
 using Core.Common.Domain.Users;
+using FunctionalTests.Commands;
+using FunctionalTests.Tests.Auctions.CreateAuction;
 using Microsoft.Extensions.DependencyInjection;
 using MongoDB.Driver;
 using System;
@@ -14,7 +16,7 @@ using Users.Domain.Repositories;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace FunctionalTests.Commands
+namespace FunctionalTests.Tests.AuctionBids.RaiseBid
 {
     [Collection(nameof(CommandTestsCollection))]
     [Trait("Category", "Functional")]
@@ -41,7 +43,7 @@ namespace FunctionalTests.Commands
             auctions.AddAuction(auction);
 
             var allAuctionBids = ServiceProvider.GetRequiredService<IAuctionBidsRepository>();
-            var auctionBids = AuctionBids.Domain.AuctionBids.CreateNew(new(auction.AggregateId), new(1, 2, 3), new(auction.Owner));
+            var auctionBids = global::AuctionBids.Domain.AuctionBids.CreateNew(new(auction.AggregateId), new(1, 2, 3), new(auction.Owner));
             ReadModelDbContext.AuctionBidsReadModel.InsertOne(new()
             {
                 AuctionId = auction.AggregateId.ToString(),
